+++ src/AntDesign.Charts/Components/Configs/ILegendMarkerStyle.cs
@@ -13,7 +13,7 @@ namespace AntDesign.Charts
     {
         public int? r { get;set; }
         public string fill { get;set; }
-        public int? fillOpacity { get;set; }
+        public decimal? fillOpacity { get;set; }
         public string stroke { get;set; }
         public int? lineWidth { get;set; }
         public int[] lineDash { get;set; }
